+++ app/controllers/external_controller.rb
@@ -1,5 +1,7 @@
 # frozen_string_literal: true
 
+require "net/http"
+
 # Handles all public facing pages.
 class ExternalController < ApplicationController
   # # GET /
@@ -44,7 +46,7 @@ class ExternalController < ApplicationController
 
   # POST /contact
   def submit_contact
-    if params[:name].present? && params[:email].present? && params[:body].present?
+    if params[:name].present? && params[:email].present? && params[:body].present? && verify_recaptcha
       UserMailer.contact(params[:name], params[:email], params[:body]).deliver_now if EMAILS_ENABLED
       redirect_to thanks_path
     else
@@ -56,4 +58,15 @@ class ExternalController < ApplicationController
   # # GET /version.json
   # def version
   # end
+
+  private
+
+  def verify_recaptcha
+    return true unless RECAPTCHA_ENABLED
+
+    uri = URI.parse("https://www.google.com/recaptcha/api/siteverify?secret=#{ENV["recaptcha_secret_key"]}&response=#{params[:recaptcha_token]}")
+    result = JSON.parse(Net::HTTP.get(uri))
+
+    result["success"] && result["score"] > 0.5 && result["action"] == "contact"
+  end
 end